DAO IMPL done

+++ src/com/cg/lms/bean/ApprovedLoans.java
@@ -4,60 +4,80 @@ public class ApprovedLoans {
 
 	int Application_Id;
 	String Customer_name;
-	int amountofloangranted;
-	int monthlyinstallment;
+	double amountofloangranted;
+	double monthlyinstallment;
 	int yearstimeperiod;
-	int downpayment;
-	int rateofinterest;
-	int totalamountpayable;
+	double downpayment;
+	double rateofinterest;
+	double totalamountpayable;
+	
+	
+	
+	
 	public int getApplication_Id() {
 		return Application_Id;
 	}
+
 	public void setApplication_Id(int application_Id) {
 		Application_Id = application_Id;
 	}
+
 	public String getCustomer_name() {
 		return Customer_name;
 	}
+
 	public void setCustomer_name(String customer_name) {
 		Customer_name = customer_name;
 	}
-	public int getAmountofloangranted() {
+
+	public double getAmountofloangranted() {
 		return amountofloangranted;
 	}
-	public void setAmountofloangranted(int amountofloangranted) {
+
+	public void setAmountofloangranted(double amountofloangranted) {
 		this.amountofloangranted = amountofloangranted;
 	}
-	public int getMonthlyinstallment() {
+
+	public double getMonthlyinstallment() {
 		return monthlyinstallment;
 	}
-	public void setMonthlyinstallment(int monthlyinstallment) {
+
+	public void setMonthlyinstallment(double monthlyinstallment) {
 		this.monthlyinstallment = monthlyinstallment;
 	}
+
 	public int getYearstimeperiod() {
 		return yearstimeperiod;
 	}
+
 	public void setYearstimeperiod(int yearstimeperiod) {
 		this.yearstimeperiod = yearstimeperiod;
 	}
-	public int getDownpayment() {
+
+	public double getDownpayment() {
 		return downpayment;
 	}
-	public void setDownpayment(int downpayment) {
+
+	public void setDownpayment(double downpayment) {
 		this.downpayment = downpayment;
 	}
-	public int getRateofinterest() {
+
+	public double getRateofinterest() {
 		return rateofinterest;
 	}
-	public void setRateofinterest(int rateofinterest) {
+
+	public void setRateofinterest(double rateofinterest) {
 		this.rateofinterest = rateofinterest;
 	}
-	public int getTotalamountpayable() {
+
+	public double getTotalamountpayable() {
 		return totalamountpayable;
 	}
-	public void setTotalamountpayable(int totalamountpayable) {
+
+	public void setTotalamountpayable(double totalamountpayable) {
 		this.totalamountpayable = totalamountpayable;
 	}
+
 	@Override
 	public String toString() {
 		return "ApprovedLoans [Application_Id=" + Application_Id
@@ -68,10 +88,11 @@ public class ApprovedLoans {
 				+ rateofinterest + ", totalamountpayable=" + totalamountpayable
 				+ "]";
 	}
+	
 	public ApprovedLoans(int application_Id, String customer_name,
-			int amountofloangranted, int monthlyinstallment,
-			int yearstimeperiod, int downpayment, int rateofinterest,
-			int totalamountpayable) {
+			double amountofloangranted, double monthlyinstallment,
+			int yearstimeperiod, double downpayment, double rateofinterest,
+			double totalamountpayable) {
 		super();
 		Application_Id = application_Id;
 		Customer_name = customer_name;
